fix(rotas): move tela do Editor para rota própria /editor
A tela do Editor de Currículo estava sobrescrevendo a rota /my-resume, que
é a rota linkada em "Meus Curriculos" na sidebar do dashboard. Agora o Editor
fica na sua própria rota /editor (mesma que aparece na topbar da própria
tela), e /my-resume volta a ser o placeholder original.
