+++ frontend/src/routes/AppRoutes.tsx
@@ -7,6 +7,7 @@ import Dashboard from "@/pages/dashboard/Dashboard";
 import { DashboardLayout } from "@/components/layout/dashboard-layout";
 import { ProtectedRoute } from "@/components/protectRouter";
 import MyResume from "@/pages/dashboard/my-resume/MyResume";
+import Editor from "@/pages/dashboard/editor/Editor";
 
 export function AppRoutes() {
   return (
@@ -16,9 +17,10 @@ export function AppRoutes() {
 
       <Route element={<ProtectedRoute />}>
         <Route path="/" element={<Home />} />
-        <Route path="/my-resume" element={<MyResume />} />
+        <Route path="/editor" element={<Editor />} />
         <Route element={<DashboardLayout />}>
           <Route path="/dashboard" element={<Dashboard />} />
+          <Route path="/my-resume" element={<MyResume />} />
         </Route>
       </Route>
     </Routes>
